+++ packages/kit/test/apps/basics/src/routes/load/layout-props/+layout.js
@@ -1,6 +1,5 @@
-/** @type {import('@sveltejs/kit').Load} */
-export async function load({ session }) {
-	session; // not necessary, but prevents the argument from being marked as unused
+/** @type {import('./$types').LayoutLoad} */
+export async function load() {
 	return {
 		// Needs to be an object, else Svelte will do by-value-comparison and skip rerender
 		obj: {}