+++ packages/kit/test/apps/basics/src/routes/store/data/only-data-changes/+page.js
@@ -1,3 +1,0 @@
-export function load({ session }) {
-	return session;
-}